[#680] Debugging get_diagnostic_equivalent_apertures)_ to be used in etendue computation
