+++ tofu/data/_class3_Aperture.py
@@ -127,3 +127,17 @@ class Aperture(_class2_Rays.Rays):
             which=which,
             key=key,
         )
+
+    def get_optics_cls(self, key=None):
+        lk = ['aperture', 'filter', 'crystal', 'grating']
+        dk = {k0: list(cls.dobj.get(k0, {}).keys()) for k0 in lk}
+
+        lok = itt.chain.from_iterable([vv for vv in dk.values()])
+        key = ds._generic_check._check_var(
+            key, 'key',
+            types=str,
+            allowed=lok,
+        )
+
+        cls = [cc for cc, lk in dk.items() if key in lk][0]
+        return key, cls
