+++ tofu/data/_class5_reflections_ptsvect.py
@@ -410,6 +410,84 @@ def _get_ptsvect(
     return ptsvect
 
 
+# ##################################################################
+# ##################################################################
+#                   preparation routine
+# ##################################################################
+
+
+def _get_project_plane(
+    plane_pt=None,
+    plane_nin=None,
+    plane_e0=None,
+    plane_e1=None,
+):
+
+    def _project_poly_on_plane_from_pt(
+        pt_x=None,
+        pt_y=None,
+        pt_z=None,
+        poly_x=None,
+        poly_y=None,
+        poly_z=None,
+        vx=None,
+        vy=None,
+        vz=None,
+        plane_pt=plane_pt,
+        plane_nin=plane_nin,
+        plane_e0=plane_e0,
+        plane_e1=plane_e1,
+    ):
+
+        sca0 = (
+            (plane_pt[0] - pt_x)*plane_nin[0]
+            + (plane_pt[1] - pt_y)*plane_nin[1]
+            + (plane_pt[2] - pt_z)*plane_nin[2]
+        )
+
+        if vx is None:
+            vx = poly_x - pt_x
+            vy = poly_y - pt_y
+            vz = poly_z - pt_z
+
+        sca1 = vx*plane_nin[0] + vy*plane_nin[1] + vz*plane_nin[2]
+
+        k = sca0 / sca1
+
+        px = pt_x + k * vx
+        py = pt_y + k * vy
+        pz = pt_z + k * vz
+
+        p0 = (
+            (px - plane_pt[0])*plane_e0[0]
+            + (py - plane_pt[1])*plane_e0[1]
+            + (pz - plane_pt[2])*plane_e0[2]
+        )
+        p1 = (
+            (px - plane_pt[0])*plane_e1[0]
+            + (py - plane_pt[1])*plane_e1[1]
+            + (pz - plane_pt[2])*plane_e1[2]
+        )
+
+        return p0, p1
+
+    def _back_to_3d(
+        x0=None,
+        x1=None,
+        plane_pt=plane_pt,
+        plane_e0=plane_e0,
+        plane_e1=plane_e1,
+    ):
+
+        return (
+            plane_pt[0] + x0*plane_e0[0] + x1*plane_e1[0],
+            plane_pt[1] + x0*plane_e0[1] + x1*plane_e1[1],
+            plane_pt[2] + x0*plane_e0[2] + x1*plane_e1[2],
+        )
+
+    return _project_poly_on_plane_from_pt, _back_to_3d
+
+
 # #################################################################
 # #################################################################
 #           Common formulas
